+++ Changelog.md
@@ -6,6 +6,9 @@
 
 - Removed deprecated `AlephAlphaClient` and `AlephAlphaModel`. Use `Client` or `AsyncClient` instead.
 - Removed deprecated `ImagePrompt`. Import `Image` instead for image prompt items.
+- New Q&A interface. We've improved the Q&A implementation, and most parameters are no longer needed.
+  - You only need to specify your documents, a query, and (optional) the max number of answers you want to receive.
+  - You no longer specify a model.
 
 ## 2.17.0